IGAPP-448: Name News and NewsList component correctly
